fix(dependencies): remove direct URL dependency for en-core-web-sm
PyPI does not allow direct URL dependencies in package metadata, which was causing
a '400 Bad Request' error during the upload process. The 'en-core-web-sm' model was
previously specified with a direct URL dependency, which is incompatible with PyPI's
requirements for package distributions.

This commit removes the direct URL for 'en-core-web-sm' from the dependencies in
'pyproject.toml'. Instead, users are encouraged to install this language model
manually or as an optional dependency documented in the README.

By addressing this, we enable the package to be published on PyPI while still
allowing users access to the necessary language models for functionality.

Further changes:
- Updated 'pyproject.toml' to reflect the adjusted dependencies without direct URLs.
- Added clarification in the README for installing 'en-core-web-sm' as a manual step if needed.
- Modified GitHub Actions workflows ('coverage.yml' and 'test-lint.yml') to ensure the
  necessary language model ('en-core-web-sm') is installed in CI/CD environments where
  it’s required for testing and coverage. This adjustment ensures automated tests continue
  to function correctly without the dependency being directly listed in the project dependencies.

This fix enhances compatibility with PyPI's publishing rules and improves the
overall user installation experience by providing clearer dependency management,
while ensuring CI/CD workflows remain functional.